Move theme_store.py and theme_manager.py to services/ (#630)
Both files contain zero Qt imports and are pure business logic
(persistence and state management with observer pattern) that
should not live in the GUI layer.

- Move theme_store.py to services/theme_store.py (JSON I/O for custom themes)
- Move theme_manager.py to services/theme_manager.py (singleton theme state)
- Keep backward-compatible re-export shims in GUI/styles/
- Update GUI/styles/__init__.py to import from services/
- Update controllers/theme_controller.py to import from services/
- Update test_theme_store.py to import from canonical location
- Add TestCanonicalLocation structural tests (5 tests)
